backend/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.20
langchain==1.1.3
langchain-openai==1.1.3
langchain-tavily==0.2.14
tavily-python==0.7.15
sqlalchemy==2.0.45
pydantic==2.12.5
pydantic-settings==2.12.0
pdfplumber==0.11.8
requests==2.32.5
beautifulsoup4==4.14.3
openai==2.11.0
python-dotenv==1.2.1
pytest==8.3.4


